refactor: introduce `helpers` folder
