+++ src/app/[locale]/_components/HomeHero/index.tsx
@@ -5,9 +5,9 @@ import { memo } from "react";
 
 import { AuthorLink, Body, Button, Cap, Heading } from "@/components/ui";
 import { routes } from "@/constants/routes";
+import { getFormattedDate } from "@/helpers/getFormattedDate";
 import { withLazyLoad } from "@/hocs/withLazyLoad";
 import { findAuthorByID } from "@/lib/authors/findAuthorByID";
-import { getFormattedDate } from "@/lib/date";
 import { findPostByID } from "@/lib/posts/findPostByID";
 
 import styles from "./HomeHero.module.scss";
