+++ src/app/[locale]/_components/PostsBlock/AllPostBlock/index.tsx
@@ -2,8 +2,8 @@ import Link from "next/link";
 
 import { AuthorLink, Body, Heading } from "@/components/ui";
 import { routes } from "@/constants/routes";
+import { getFormattedDate } from "@/helpers/getFormattedDate";
 import { findAuthorByID } from "@/lib/authors/findAuthorByID";
-import { getFormattedDate } from "@/lib/date";
 import { findPostByID } from "@/lib/posts/findPostByID";
 
 import styles from "./AllPostBlock.module.scss";
