+++ src/app/[locale]/_components/layout/LayoutFooter/index.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 import { Body, Footer, Heading } from "@/components/ui";
 import { address, emailAndPhone } from "@/constants/contactInfo";
 import { socialLinks } from "@/constants/links";
-import { getLinks } from "@/lib/links";
+import { getLinks } from "@/helpers/getLinks";
 
 import {
   LocaleSwitcherSkeleton,
