+++ src/app/[locale]/blog/[id]/_components/PostDetails/index.tsx
@@ -3,9 +3,9 @@ import Link from "next/link";
 
 import { Avatar, Body, Heading } from "@/components/ui";
 import { routes } from "@/constants/routes";
+import { getFormattedDate } from "@/helpers/getFormattedDate";
 import { findAuthorByID } from "@/lib/authors/findAuthorByID";
 import { findCategoryByID } from "@/lib/categories/findCategoryByID";
-import { getFormattedDate } from "@/lib/date";
 
 import { type PostDetailsProps } from "./interfaces";
 import styles from "./PostDetails.module.scss";
