+++ src/helpers/isSelectedOptionDisabled.ts
